# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/pagemem/wal/record/delta/DataPageInsertRecord.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheMapEntry.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheTtlManager.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/IgniteCacheOffheapManager.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/IgniteCacheOffheapManagerImpl.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/database/freelist/FreeList.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/cache/database/tree/io/DataPageIO.java
#	modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/opt/GridH2Row.java

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/IgniteCacheOffheapManagerImpl.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.processors.cache;
 
 import java.nio.ByteBuffer;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.Set;
@@ -34,6 +33,7 @@ import org.apache.ignite.internal.pagemem.PageIdUtils;
 import org.apache.ignite.internal.pagemem.PageMemory;
 import org.apache.ignite.internal.processors.affinity.AffinityTopologyVersion;
 import org.apache.ignite.internal.processors.cache.database.CacheDataRow;
+import org.apache.ignite.internal.processors.cache.database.CacheDataRowAdapter;
 import org.apache.ignite.internal.processors.cache.database.IgniteCacheDatabaseSharedManager;
 import org.apache.ignite.internal.processors.cache.database.MetaStore;
 import org.apache.ignite.internal.processors.cache.database.MetadataStorage;
@@ -44,7 +44,6 @@ import org.apache.ignite.internal.processors.cache.database.tree.BPlusTree;
 import org.apache.ignite.internal.processors.cache.database.tree.io.BPlusIO;
 import org.apache.ignite.internal.processors.cache.database.tree.io.BPlusInnerIO;
 import org.apache.ignite.internal.processors.cache.database.tree.io.BPlusLeafIO;
-import org.apache.ignite.internal.processors.cache.database.tree.io.DataPageIO;
 import org.apache.ignite.internal.processors.cache.database.tree.io.IOVersions;
 import org.apache.ignite.internal.processors.cache.database.tree.reuse.ReuseList;
 import org.apache.ignite.internal.processors.cache.distributed.dht.GridDhtInvalidPartitionException;
@@ -55,19 +54,17 @@ import org.apache.ignite.internal.processors.cache.version.GridCacheVersion;
 import org.apache.ignite.internal.processors.query.GridQueryProcessor;
 import org.apache.ignite.internal.util.GridCloseableIteratorAdapter;
 import org.apache.ignite.internal.util.GridEmptyCloseableIterator;
+import org.apache.ignite.internal.util.GridLongList;
 import org.apache.ignite.internal.util.lang.GridCloseableIterator;
 import org.apache.ignite.internal.util.lang.GridCursor;
 import org.apache.ignite.internal.util.lang.GridIterator;
 import org.apache.ignite.internal.util.typedef.F;
-import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.lang.IgniteClosure;
 import org.apache.ignite.lang.IgnitePredicate;
 import org.jetbrains.annotations.Nullable;
 
 import static org.apache.ignite.internal.IgniteComponentType.INDEXING;
-import static org.apache.ignite.internal.pagemem.PageIdUtils.dwordsOffset;
-import static org.apache.ignite.internal.pagemem.PageIdUtils.pageId;
 
 /**
  *
@@ -173,7 +170,7 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
     @Override protected void stop0(final boolean cancel, final boolean destroy) {
         super.stop0(cancel, destroy);
 
-        if (destroy) {
+        if (destroy && !cctx.kernalContext().clientNode()) {
             destroyCacheDataStructures();
 
             cctx.shared().database().pageMemory().clear(cctx.cacheId());
@@ -186,47 +183,49 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
     protected void destroyCacheDataStructures() {
         final PageMemory pageMem = cctx.shared().database().pageMemory();
 
-        if (!cctx.shared().database().persistenceEnabled()) {
-            try (Page meta = pageMem.metaPage(cctx.cacheId())) {
-                final ByteBuffer buf = meta.getForWrite();
-
-                try {
-                    // Set number of initialized pages to 0.
-                    buf.putInt(0);
-                }
-                finally {
-                    meta.releaseWrite(true);
-                }
-            }
-            catch (IgniteCheckedException e) {
-                throw new IgniteException(e.getMessage(), e);
-            }
+        try (Page meta = pageMem.metaPage(cctx.cacheId())) {
+            final ByteBuffer buf = meta.getForWrite();
 
             try {
-                if (locCacheDataStore != null)
-                    locCacheDataStore.destroy();
-
-                for (CacheDataStore store : partDataStores.values()) {
-                    store.destroy();
-                }
-
-                metaStore.destroy();
-
-                Collection<Long> pages = freeList.pages();
-
-                pages.addAll(reuseList.pages());
-
-                reuseList.destroy();
-
-                freeList.destroy();
-
-                for (Long pageId : pages) {
-                    pageMem.freePage(cctx.cacheId(), pageId);
-                }
+                // Set number of initialized pages to 0.
+                buf.putInt(0);
             }
-            catch (IgniteCheckedException e) {
-                throw new IgniteException(e.getMessage(), e);
+            finally {
+                meta.releaseWrite(true);
             }
+        }
+        catch (IgniteCheckedException e) {
+            throw new IgniteException(e.getMessage(), e);
+        }
+
+        try {
+            if (locCacheDataStore != null)
+                locCacheDataStore.destroy();
+
+            for (CacheDataStore store : partDataStores.values()) {
+                store.destroy();
+            }
+
+            metaStore.destroy();
+
+            GridLongList pagesList = new GridLongList();
+
+            freeList.pages(pagesList);
+
+            reuseList.pages(pagesList);
+
+            reuseList.destroy();
+
+            freeList.destroy();
+
+            for (int i = 0; i < pagesList.size(); i++) {
+                long pageId = pagesList.get(i);
+
+                pageMem.freePage(cctx.cacheId(), pageId);
+            }
+        }
+        catch (IgniteCheckedException e) {
+            throw new IgniteException(e.getMessage(), e);
         }
     }
 
@@ -376,21 +375,13 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
 
     /** {@inheritDoc} */
     @SuppressWarnings("unchecked")
-    @Nullable public CacheDataRow read(GridCacheMapEntry entry)
+    @Override @Nullable public CacheDataRow read(GridCacheMapEntry entry)
         throws IgniteCheckedException {
-        try {
-            KeyCacheObject key = entry.key();
+        KeyCacheObject key = entry.key();
 
-            assert cctx.isLocal() || entry.localPartition() != null : entry;
+        assert cctx.isLocal() || entry.localPartition() != null : entry;
 
-            return dataStore(entry.localPartition()).find(key);
-        }
-        catch (IgniteCheckedException e) {
-            throw e;
-        }
-        catch (Exception e) {
-            throw new IgniteCheckedException("Failed to read entry: " + entry.key(), e);
-        }
+        return dataStore(entry.localPartition()).find(key);
     }
 
     /** {@inheritDoc} */
@@ -641,15 +632,6 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         };
     }
 
-    /**
-     * @param pageId Page ID.
-     * @return Page.
-     * @throws IgniteCheckedException If failed.
-     */
-    private Page page(long pageId) throws IgniteCheckedException {
-        return cctx.shared().database().pageMemory().page(cctx.cacheId(), pageId);
-    }
-
     /** {@inheritDoc} */
     @Override public final CacheDataStore createCacheDataStore(int p,
         CacheDataStore.Listener lsnr) throws IgniteCheckedException {
@@ -674,6 +656,20 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         partDataStores.put(p, dataStore);
 
         return dataStore;
+    }
+
+    /** {@inheritDoc} */
+    @Override public void destroyCacheDataStore(int p, CacheDataStore store) throws IgniteCheckedException {
+        try {
+            partDataStores.remove(p, store);
+
+            store.destroy();
+
+            meta().dropRootPage(store.name());
+        }
+        catch (IgniteCheckedException e) {
+            throw new IgniteException(e);
+        }
     }
 
     /**
@@ -749,6 +745,11 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         }
 
         /** {@inheritDoc} */
+        @Override public String name() {
+            return name;
+        }
+
+        /** {@inheritDoc} */
         @Override public void update(KeyCacheObject key,
             int p,
             CacheObject val,
@@ -758,7 +759,7 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
 
             rowStore.addRow(dataRow);
 
-            assert dataRow.link != 0 : dataRow;
+            assert dataRow.link() != 0 : dataRow;
 
             DataRow old = dataTree.put(dataRow);
 
@@ -767,11 +768,11 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
 
                 assert qryMgr.enabled();
 
-                qryMgr.store(key, p, val, ver, expireTime, dataRow.link);
+                qryMgr.store(key, p, val, ver, expireTime, dataRow.link());
             }
 
             if (old != null) {
-                assert old.link != 0 : old;
+                assert old.link() != 0 : old;
 
                 if (pendingEntries != null && old.expTime != 0) {
                     PendingRow rmvd = pendingEntries.remove(new PendingRow(old.expTime, old.link, false));
@@ -779,7 +780,7 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
                     assert rmvd != null;
                 }
 
-                rowStore.removeRow(old.link);
+                rowStore.removeRow(old.link());
             }
             else
                 lsnr.onInsert();
@@ -793,11 +794,6 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
             CacheObject prevVal,
             GridCacheVersion prevVer,
             int partId) throws IgniteCheckedException {
-            DataRow dataRow = dataTree.remove(new KeySearchRow(key.hashCode(), key, 0));
-
-            if (dataRow == null)
-                return;
-
             if (indexingEnabled) {
                 GridCacheQueryManager qryMgr = cctx.queries();
 
@@ -806,17 +802,21 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
                 qryMgr.remove(key, partId, prevVal, prevVer);
             }
 
-            assert dataRow.link != 0 : dataRow;
+            DataRow dataRow = dataTree.remove(new KeySearchRow(key.hashCode(), key, 0));
 
-            if (pendingEntries != null && dataRow.expTime != 0) {
-                PendingRow rmvd = pendingEntries.remove(new PendingRow(dataRow.expTime, dataRow.link, false));
+            if (dataRow != null) {
+                assert dataRow.link() != 0 : dataRow;
 
-                assert rmvd != null;
+                if (pendingEntries != null && dataRow.expTime != 0) {
+                    PendingRow rmvd = pendingEntries.remove(new PendingRow(dataRow.expTime, dataRow.link, false));
+
+                    assert rmvd != null;
+                }
+
+                rowStore.removeRow(dataRow.link());
+
+                lsnr.onRemove();
             }
-
-            rowStore.removeRow(dataRow.link);
-
-            lsnr.onRemove();
         }
 
         /** {@inheritDoc} */
@@ -842,15 +842,9 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
     /**
      *
      */
-    private class KeySearchRow {
+    private class KeySearchRow extends CacheDataRowAdapter {
         /** */
         protected int hash;
-
-        /** */
-        protected KeyCacheObject key;
-
-        /** */
-        protected long link;
 
         /**
          * @param hash Hash code.
@@ -858,49 +852,28 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
          * @param link Link.
          */
         KeySearchRow(int hash, KeyCacheObject key, long link) {
-            this.hash = hash;
-            this.key = key;
-            this.link = link;
-        }
+            super(link);
 
-        /**
-         * @param buf Buffer.
-         * @throws IgniteCheckedException If failed.
-         */
-        protected void doInitData(ByteBuffer buf) throws IgniteCheckedException {
-            key = cctx.cacheObjects().toKeyCacheObject(cctx.cacheObjectContext(), buf);
+            this.key = key;
+            this.hash = hash;
         }
 
         /**
          * Init data.
+         *
+         * @param keyOnly Initialize only key.
          */
-        protected final void initData() {
+        protected final void initData(boolean keyOnly) {
             if (key != null)
                 return;
 
-            assert link != 0;
+            assert link() != 0;
 
-            try (Page page = page(pageId(link))) {
-                ByteBuffer buf = page.getForRead();
-
-                try {
-                    DataPageIO io = DataPageIO.VERSIONS.forPage(buf);
-
-                    int dataOff = io.getDataOffset(buf, dwordsOffset(link));
-
-                    buf.position(dataOff);
-
-                    // Skip entry size.
-                    buf.getShort();
-
-                    doInitData(buf);
-                }
-                finally {
-                    page.releaseRead();
-                }
+            try {
+                initFromLink(cctx, keyOnly);
             }
             catch (IgniteCheckedException e) {
-                throw new IgniteException(e);
+                throw new IgniteException(e.getMessage(), e);
             }
         }
 
@@ -908,29 +881,18 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
          * @return Key.
          */
         public KeyCacheObject key() {
-            initData();
+            initData(true);
 
             return key;
-        }
-
-        /** {@inheritDoc} */
-        public String toString() {
-            return S.toString(KeySearchRow.class, this);
         }
     }
 
     /**
      *
      */
-    class DataRow extends KeySearchRow implements CacheDataRow {
+    private class DataRow extends KeySearchRow {
         /** */
-        private CacheObject val;
-
-        /** */
-        private GridCacheVersion ver;
-
-        /** */
-        private int part = -1;
+        int part = -1;
 
         /** */
         private long expTime;
@@ -953,7 +915,7 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
             part = PageIdUtils.partId(link);
 
             // We can not init data row lazily because underlying buffer can be concurrently cleared.
-            initData();
+            initData(false);
         }
 
         /**
@@ -973,55 +935,13 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         }
 
         /** {@inheritDoc} */
-        @Override protected void doInitData(ByteBuffer buf) throws IgniteCheckedException {
-            key = cctx.cacheObjects().toKeyCacheObject(cctx.cacheObjectContext(), buf);
-            val = cctx.cacheObjects().toCacheObject(cctx.cacheObjectContext(), buf);
-
-            int topVer = buf.getInt();
-            int nodeOrderDrId = buf.getInt();
-            long globalTime = buf.getLong();
-            long order = buf.getLong();
-
-            ver = new GridCacheVersion(topVer, nodeOrderDrId, globalTime, order);
-            expTime = buf.getLong();
-        }
-
-        /** {@inheritDoc} */
-        @Override public CacheObject value() {
-            assert val != null;
-
-            return val;
-        }
-
-        /** {@inheritDoc} */
-        @Override public GridCacheVersion version() {
-            assert ver != null;
-
-            return ver;
-        }
-
-        /** {@inheritDoc} */
         @Override public int partition() {
-            assert part != -1;
-
             return part;
-        }
-
-        /** {@inheritDoc} */
-        @Override public long link() {
-            return link;
         }
 
         /** {@inheritDoc} */
         @Override public void link(long link) {
             this.link = link;
-        }
-
-        /** {@inheritDoc} */
-        @Override public long expireTime() {
-            initData();
-
-            return expTime;
         }
 
         /** {@inheritDoc} */
@@ -1074,12 +994,14 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         /** {@inheritDoc} */
         @Override protected int compare(BPlusIO<KeySearchRow> io, ByteBuffer buf, int idx, KeySearchRow row)
             throws IgniteCheckedException {
-            KeySearchRow row0 = io.getLookupRow(this, buf, idx);
+            int hash = ((RowLinkIO)io).getHash(buf, idx);
 
-            int cmp = Integer.compare(row0.hash, row.hash);
+            int cmp = Integer.compare(hash, row.hash);
 
             if (cmp != 0)
                 return cmp;
+
+            KeySearchRow row0 = io.getLookupRow(this, buf, idx);
 
             return compareKeys(row0.key(), row.key());
         }
@@ -1196,9 +1118,9 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
 
         /** {@inheritDoc} */
         @Override public void storeByOffset(ByteBuffer buf, int off, KeySearchRow row) {
-            assert row.link != 0;
+            assert row.link() != 0;
 
-            store0(buf, off, row.link, row.hash);
+            store0(buf, off, row.link(), row.hash);
         }
 
         /** {@inheritDoc} */
@@ -1251,9 +1173,9 @@ public class IgniteCacheOffheapManagerImpl extends GridCacheManagerAdapter imple
         @Override public void storeByOffset(ByteBuffer buf, int off, KeySearchRow row) {
             DataRow row0 = (DataRow)row;
 
-            assert row0.link != 0;
+            assert row0.link() != 0;
 
-            store0(buf, off, row.link, row.hash);
+            store0(buf, off, row.link(), row.hash);
         }
 
         /** {@inheritDoc} */
